+++ src/renderer.h
@@ -18,13 +18,13 @@ class Renderer {
         {}
 
         int16_t translateX;
-        int16_t translateY;
+        int8_t translateY;
 
         void fillRect(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t color = WHITE);
         void drawRect(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t color = WHITE);
         void drawOverwrite(int16_t x, int16_t y, const uint8_t* bitmap, uint8_t frame, MirrorMode mirror = 0, DrawMode drawMode = Normal);
         void drawPlusMask(int16_t x, int16_t y, const uint8_t* bitmap, uint8_t frame, MirrorMode mirror = 0, DrawMode drawMode = Normal);
-        void drawString(int16_t x, int16_t y, const uint8_t* str);
+        void drawString(int8_t x, int8_t y, const uint8_t* str);
 };
 
 #endif
